refactor conversations and fix some group bugs
