+++ src/open_llm_vtuber/conversations/types.py
@@ -1,0 +1,48 @@
+from typing import List, Dict, Union, Any, Callable, Optional, TypedDict, Awaitable
+from dataclasses import dataclass
+from pydantic import BaseModel
+
+from ..agent.output_types import Actions, DisplayText
+
+# Type definitions
+WebSocketSend = Callable[[str], Awaitable[None]]
+BroadcastFunc = Callable[[List[str], dict, Optional[str]], Awaitable[None]]
+
+
+class AudioPayload(TypedDict):
+    """Type definition for audio payload"""
+
+    type: str
+    audio: Optional[str]
+    volumes: Optional[List[float]]
+    slice_length: Optional[int]
+    display_text: Optional[DisplayText]
+    actions: Optional[Actions]
+    forwarded: Optional[bool]
+
+
+@dataclass
+class BroadcastContext:
+    """Context for broadcasting messages in group chat"""
+
+    broadcast_func: Optional[BroadcastFunc] = None
+    group_members: Optional[List[str]] = None
+    current_client_uid: Optional[str] = None
+
+
+class ConversationConfig(BaseModel):
+    """Configuration for conversation chain"""
+
+    conf_uid: str = ""
+    history_uid: str = ""
+    client_uid: str = ""
+    character_name: str = "AI"
+
+
+class GroupConversationState(BaseModel):
+    """State for group conversation"""
+
+    conversation_history: List[str] = []
+    memory_index: Dict[str, int] = {}
+    group_queue: List[str] = []
+    session_emoji: str = ""
