+++ src/open_llm_vtuber/service_context.py
@@ -186,7 +186,10 @@ class ServiceContext:
             return
 
         system_prompt = self.construct_system_prompt(persona_prompt)
-
+        
+        # Pass avatar to agent factory
+        avatar = self.character_config.avatar or ""  # Get avatar from config
+        
         try:
             self.agent_engine = AgentFactory.create_agent(
                 conversation_agent_choice=agent_config.conversation_agent_choice,
@@ -195,6 +198,7 @@ class ServiceContext:
                 system_prompt=system_prompt,
                 live2d_model=self.live2d_model,
                 tts_preprocessor_config=self.character_config.tts_preprocessor_config,
+                character_avatar=avatar  # Add avatar parameter
             )
 
             logger.debug(f"Agent choice: {agent_config.conversation_agent_choice}")
